Schedule folder, view schedule, add course updated
Added a folder to hold csv files of schedules. They're saved as studentID_schedule.csv. System reads in schedule when student's log in
Added view schedule option for students, prints their current schedule.
Updated addCourse to add the course to the student's schedule and write it to their schedule.csv